requirements.txt
openai==1.50.2
python-dotenv